+++ draft-packages/popover/KaizenDraft/Popover/PopoverLegacy.tsx
@@ -66,13 +66,11 @@ const mapArrowSideToClass = (side: Side): string => {
   }
 }
 
-type PopoverLegacy = React.FunctionComponent<PopoverLegacyProps>
-
 /**
  * Please use `usePopover` or `Popover` instead
  * @deprecated
  */
-export const PopoverLegacy: PopoverLegacy = React.forwardRef<
+export const PopoverLegacy = React.forwardRef<
   HTMLDivElement,
   PopoverLegacyProps
 >(