+++ packages/button/src/Button/PaginationLink.tsx
@@ -12,9 +12,7 @@ export interface PaginationLinkProps extends GenericProps {
  * {@link https://cultureamp.design/components/pagination/ Guidance} |
  * {@link https://cultureamp.design/storybook/?path=/story/components-button-pagination-buttons--default-kaizen-pagination-link Storybook}
  */
-const PaginationLink: React.VFC<PaginationLinkProps> = (
-  props: PaginationLinkProps
-) => (
+const PaginationLink = (props: PaginationLinkProps): JSX.Element => (
   <GenericButton
     {...props}
     paginationLink