+++ packages/pagination/src/TruncateIndicator.tsx
@@ -3,7 +3,7 @@ import { Icon } from "@kaizen/component-library"
 import ellipsis from "@kaizen/component-library/icons/ellipsis.icon.svg"
 import styles from "./Pagination.module.scss"
 
-export const TruncateIndicator = () => (
+export const TruncateIndicator = (): JSX.Element => (
   <div
     className={styles.truncateIndicatorWrapper}
     data-testid="truncate-indicator"