+++ packages/rich-text-editor/src/EditableRichTextContent/EditableRichTextContent.tsx
@@ -17,16 +17,16 @@ export interface EditableRichTextContentProps
 const handleEditableClick = (
   e: MouseEvent<HTMLElement>,
   onClick: (event: MouseEvent<HTMLElement>) => void
-) => {
+): void => {
   if ((e.target as HTMLElement).hasAttribute("href")) {
     return
   }
   return onClick(e)
 }
 
-export const EditableRichTextContent: React.VFC<
-  EditableRichTextContentProps
-> = props => {
+export const EditableRichTextContent = (
+  props: EditableRichTextContentProps
+): JSX.Element => {
   const {
     onClick,
     content,
@@ -42,7 +42,7 @@ export const EditableRichTextContent: React.VFC<
       {/* Disabling these a11y linting errors because there is a <button> that mitigates these concerns. The onClick here is just an additional layer. */}
       {/* eslint-disable-next-line jsx-a11y/no-static-element-interactions, jsx-a11y/click-events-have-key-events */}
       <div
-        onClick={e => handleEditableClick(e, onClick)}
+        onClick={(e): void => handleEditableClick(e, onClick)}
         className={classnames(styles.editableContainer, classNameOverride)}
         {...restProps}
       >