+++ draft-packages/empty-state/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.3.15](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-empty-state@5.3.14...@kaizen/draft-empty-state@5.3.15) (2023-01-04)
+
+**Note:** Version bump only for package @kaizen/draft-empty-state
+
+
+
+
+
 ## [5.3.14](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-empty-state@5.3.13...@kaizen/draft-empty-state@5.3.14) (2023-01-03)
 
 **Note:** Version bump only for package @kaizen/draft-empty-state