+++ packages/design-tokens/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.1.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/design-tokens@10.1.1...@kaizen/design-tokens@10.1.2) (2023-01-04)
+
+**Note:** Version bump only for package @kaizen/design-tokens
+
+
+
+
+
+## [10.1.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/design-tokens@10.1.0...@kaizen/design-tokens@10.1.1) (2023-01-04)
+
+**Note:** Version bump only for package @kaizen/design-tokens
+
+
+
+
+
 # [10.1.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/design-tokens@10.0.14...@kaizen/design-tokens@10.1.0) (2022-12-08)