+++ packages/select/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.3.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/select@6.3.2...@kaizen/select@6.3.3) (2023-01-04)
+
+**Note:** Version bump only for package @kaizen/select
+
+
+
+
+
+## [6.3.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/select@6.3.1...@kaizen/select@6.3.2) (2023-01-04)
+
+**Note:** Version bump only for package @kaizen/select
+
+
+
+
+
 ## [6.3.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/select@6.3.0...@kaizen/select@6.3.1) (2023-01-03)
 
 **Note:** Version bump only for package @kaizen/select